+++ src/pages/mypage/withdrawal/fail.page.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image';
 import Link from 'next/link';
 import { button } from '@components/common/Button/styles';
-import { CommonNoticePage } from '@pages/login/index.page';
+import { CommonNoticePage } from '@components/Layout/NoticePageLayout';
 
 export default function Fail() {
   return (
